refactor: rm api_client
