+++ inky-mbta-tracker/mbta_client_extended.py
@@ -603,3 +603,35 @@ async def watch_alerts(
             session=session,
             transit_time_min=0,
         )
+
+
+def get_default_routes() -> List[str]:
+    """Get the default list of commuter rail routes for precaching."""
+    return [
+        "CR-Worcester",
+        "CR-Framingham",
+        "CR-Franklin",
+        "CR-Foxboro",
+        "CR-Providence",
+        "CR-Stoughton",
+        "CR-Needham",
+        "CR-Fairmount",
+        "CR-Fitchburg",
+        "CR-Lowell",
+        "CR-Haverhill",
+        "CR-Newburyport",
+        "CR-Rockport",
+        "CR-Kingston",
+        "CR-Plymouth",
+        "CR-Greenbush",
+    ]
+
+
+def get_default_target_stations() -> List[str]:
+    """Get the default list of stations for precaching."""
+    return [
+        "place-sstat",  # South Station
+        "place-north",  # North Station
+        "place-bbsta",  # Back Bay
+        "place-rugg",  # Ruggles
+    ]
